+++ template/src/framework/react/Counter.tsx
@@ -1,3 +1,4 @@
+import { useStore } from "@nanostores/react";
 import { useState } from "react";
 
 import AstroLogo from "../../images/astro.png?url";
@@ -21,18 +22,49 @@ export default function Counter({
     setCount((i) => setCounterCount(i, Operation.Subtract));
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const showParentMessage = () => (window as any)[showMessage](count);
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const nanoStoreValue = useStore((window as any).Stores["reactStore"]);
 
   return (
     <>
       {header}
-      <div className="counter-controls">
-        <button onClick={subtract}>-</button>
-        <pre>{count}</pre>
-        <button onClick={add}>+</button>
-      </div>
-      <div>{children}</div>
-      <div className="counter-message">
-        <button onClick={showParentMessage}>Send value</button>
+      <div className="card-grid">
+        <div className="card">
+          Internal counter controls. It keeps state within the component.
+          <div className="card-content">
+            <div className="counter-controls">
+              <button onClick={subtract}>-</button>
+              <pre>{count}</pre>
+              <button onClick={add}>+</button>
+            </div>
+          </div>
+        </div>
+        <div className="card">
+          The button sends the current count value to a function in the parent
+          component.
+          <div className="card-content">
+            <div>
+              <button className="card-btn" onClick={showParentMessage}>
+                Send value
+              </button>
+            </div>
+          </div>
+        </div>
+        <div className="card">
+          Slot content coming in to the component
+          <div className="card-content">
+            <div>{children}</div>
+          </div>
+        </div>
+        <div className="card">
+          Nano Store content
+          <div className="card-content">
+            <div>
+              <strong>Nano Store value:</strong>
+              <div id="nanostore">{nanoStoreValue}</div>
+            </div>
+          </div>
+        </div>
       </div>
       <div className="counter-logos">
         <img alt="OutSystems logo" src={OutSystemsLogo} />