+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Added Nano Stores for React and Vue.
+
 ### Changed
 - Updated docs dependency typescript-eslint to 8.54.0.
 - Updated docs dependency starlight-versions to 0.6.0.
@@ -17,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated docs dependency @eslint/compat to 2.0.2.
 - Updated docs dependency @astrojs/starlight to 0.37.5.
 - Updated Deno allowed scripts.
-- Added template peer dependencies @angular-devkit/architect, @angular/cli, hono and vite.
+- Added template peer dependencies @angular-devkit/architect, @angular/cli, hono, vite and zod.
 - Updated template dependency vue 3.25.7.
 - Updated template dependency vitest to 4.0.18.
 - Updated template dependency typescript-eslint to 8.54.0.